feat: working websocket

+++ backend/src/main/java/stud/ntnu/krisefikser/map/service/EventService.java
@@ -99,9 +99,10 @@ public class EventService {
    * @throws EntityNotFoundException If no event with the specified ID exists
    */
   public void deleteEvent(Long id) {
-    Event event = eventRepository.findById(id)
-        .orElseThrow(() -> new EntityNotFoundException("Event not found with id: " + id));
-    eventWebSocketService.notifyEventDeletion(event);
+    if (!eventRepository.existsById(id)) {
+      throw new EntityNotFoundException("Event not found with id: " + id);
+    }
+    eventWebSocketService.notifyEventDeletion(id);
     eventRepository.deleteById(id);
   }
 }
